Split RadioControl component
